production testing 3

+++ testMaze2/Dot.swift
@@ -19,7 +19,7 @@ class Dot: UIView {
     var tap = UITapGestureRecognizer()
     var tapDelegate:DotTap?
     var shapeLayer = CAShapeLayer()
-    init(size: CGSize = CGSize(width: 45*UIScreen.main.bounds.width/375, height: 45*UIScreen.main.bounds.width/375), color: UIColor, origin: CGPoint, colorScheme: ColorScheme) {
+    init(size: CGSize = CGSize(width: CGFloat(Int(45*UIScreen.main.bounds.width/375)), height: CGFloat(Int(45*UIScreen.main.bounds.width/375))), color: UIColor, origin: CGPoint, colorScheme: ColorScheme) {
         super.init(frame: CGRect(origin: origin, size: size))
         dotColor = color
         tap = UITapGestureRecognizer(target: self, action: #selector(Dot.tapFunc(_:)))
@@ -28,7 +28,7 @@ class Dot: UIView {
         self.addGestureRecognizer(tap)
         
         print("dotframe: \(self.frame), scheme \(_colorScheme!)")
-        let circlePath = UIBezierPath(arcCenter: CGPoint(x: 22.5*UIScreen.main.bounds.width/375, y: 22.5*UIScreen.main.bounds.width/375 ), radius: 15*UIScreen.main.bounds.width/375, startAngle: 0, endAngle:2 * .pi, clockwise: true)
+        let circlePath = UIBezierPath(arcCenter: CGPoint(x: CGFloat(Int(22.5*UIScreen.main.bounds.width/375)), y: CGFloat(Int(22.5*UIScreen.main.bounds.width/375)) ), radius: CGFloat(Int(15*UIScreen.main.bounds.width/375)), startAngle: 0, endAngle:2 * .pi, clockwise: true)
         
         
         shapeLayer.path = circlePath.cgPath
